Fix mypy type errors in AI modules
- embeddings.py: Cast cosine_similarity return to float
- llm_adapter.py: Wrap json.loads returns in dict() for type safety
- citation_check.py, deep_dive.py, lineage.py, update_detection.py:
  Validate json.loads returns dict before returning

+++ curious_now/ai/embeddings.py
@@ -416,7 +416,7 @@ def cosine_similarity(embedding1: list[float], embedding2: list[float]) -> float
     if magnitude1 == 0 or magnitude2 == 0:
         return 0.0
 
-    return dot_product / (magnitude1 * magnitude2)
+    return float(dot_product / (magnitude1 * magnitude2))
 
 
 async def generate_embeddings_batch(
